move random_shuffle and random_sample logic to utils
