+++ salt/modules/mod_random.py
@@ -10,9 +10,9 @@ import base64
 import hashlib
 import random
 
+import salt.utils.data
 import salt.utils.pycrypto
 from salt.exceptions import SaltInvocationError
-from salt.utils.decorators.jinja import jinja_filter
 
 ALGORITHMS_ATTR_NAME = "algorithms_guaranteed"
 
@@ -280,7 +280,6 @@ def seed(range=10, hash=None):
     return random.randrange(range)
 
 
-@jinja_filter("random_sample")
 def sample(value, size, seed=None):
     """
     Return a given sample size from a list. By default, the random number
@@ -303,14 +302,9 @@ def sample(value, size, seed=None):
 
         salt '*' random.sample '["one", "two"]' 1 seed="something"
     """
-    if seed is None:
-        ret = random.sample(value, size)
-    else:
-        ret = random.Random(hash(seed)).sample(value, size)
-    return ret
+    return salt.utils.data.sample(value, size, seed=seed)
 
 
-@jinja_filter("random_shuffle")
 def shuffle(value, seed=None):
     """
     Return a shuffled copy of an input list. By default, the random number
@@ -330,4 +324,4 @@ def shuffle(value, seed=None):
 
         salt '*' random.shuffle '["one", "two"]' seed="something"
     """
-    return sample(value, len(value), seed=seed)
+    return salt.utils.data.shuffle(value, seed=seed)
