+++ tests/pytests/unit/utils/test_data.py
@@ -81,3 +81,22 @@ def test_flatten_recursion_error():
     with pytest.raises(RecursionError) as err:
         salt.utils.data.flatten(data)
     assert str(err.value) == "Reference cycle detected. Check input list."
+
+
+def test_sample():
+    lst = ["one", "two", "three", "four"]
+    assert len(salt.utils.data.sample(lst, 0)) == 0
+    assert len(salt.utils.data.sample(lst, 2)) == 2
+    pytest.raises(ValueError, salt.utils.data.sample, lst, 5)
+    assert salt.utils.data.sample(lst, 2, seed="static") == ["four", "two"]
+
+
+def test_shuffle():
+    lst = ["one", "two", "three", "four"]
+    assert len(salt.utils.data.shuffle(lst)) == 4
+    assert salt.utils.data.shuffle(lst, seed="static") == [
+        "four",
+        "two",
+        "three",
+        "one",
+    ]
